+++ lib/src/client_io.dart
@@ -58,7 +58,7 @@ class ClientIO extends ClientBase with ClientMixin {
       'x-sdk-name': 'Flutter',
       'x-sdk-platform': 'client',
       'x-sdk-language': 'flutter',
-      'x-sdk-version': '24.2.0',
+      'x-sdk-version': '25.0.0',
       'X-Appwrite-Response-Format': '1.9.5',
     };